requirements.txt
fastapi==0.103.0
psycopg2-binary==2.9.7
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
SQLAlchemy==2.0.20
uvicorn==0.23.2
python-dotenv==1.0.0